requirements.txt
aioboto3==11.2.0
aiobotocore==2.5.0
aiohttp==3.8.4
aioitertools==0.11.0
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
black==23.3.0
bleach==6.0.0
boto3==1.26.76
botocore==1.29.76
brotlipy==0.7.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cryptography==40.0.2
cycler==0.11.0
deeplake==3.6.10
dill==0.3.6
docutils==0.20.1
entrypoints==0.4
filelock==3.12.0
fonttools==4.39.4
frozenlist==1.4.0
gensim==4.3.1
gmpy2==2.1.2
humbug==0.3.2
idna==3.4
imageio==2.31.1
importlib-metadata==6.6.0
importlib-resources==5.12.0
jaraco.classes==3.2.3
jeepney==0.8.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
keyring==23.13.1
kiwisolver==1.4.4
lazy_loader==0.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplot==0.1.9
matplotlib==3.7.1
mdurl==0.1.2
more-itertools==9.1.0
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
mypy-extensions==1.0.0
nest-asyncio==1.5.6
networkx==3.1
numcodecs==0.11.0
numpy==1.24.3
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.1
pathos==0.3.0
pathspec==0.11.1
Pillow==9.5.0
pip==23.1.2
pkginfo==1.9.6
platformdirs==3.5.1
pox==0.3.2
ppft==1.7.6.6
pycparser==2.21
Pygments==2.15.1
PyJWT==2.7.0
pyloco==0.0.139
pyOpenSSL==23.1.1
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
readme-renderer==37.3
requests==2.29.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.3.5
s3transfer==0.6.1
scikit-image==0.21.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
SecretStorage==3.3.3
setuptools==67.7.2
SimpleWebSocketServer==0.1.2
six==1.16.0
sklearn==0.0.post5
smart-open==6.3.0
sympy==1.12
threadpoolctl==3.1.0
tifffile==2023.7.10
tomli==2.0.1
torch==2.0.0.post200
torchvision==0.15.1a0+60a3e72
tqdm==4.65.0
tsnecuda==3.0.1
twine==4.0.2
typing==3.7.4.3
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
ushlex==0.99.1
webencodings==0.5.1
websocket-client==1.5.1
wheel==0.40.0
wrapt==1.15.0
xgboost==1.7.5
yarl==1.9.2
zipp==3.15.0
